pyproject.toml
[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[project]
name = "finance-research-data"
version = "0.1.0"
description = "A Python package that downloads raw data commonly used in finance academic research."
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
authors = [{ name = "Mihai Ion" }]

# Runtime dependencies
dependencies = [
    "requests>=2.32",
    "pandas>=2.1",
    "python-dotenv>=1.0",
    "wrds>=3.1",
    "openpyxl>=3.1.2"
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-cov>=5",
    "ruff>=0.5",
    "mypy>=1.10",
    "build>=1.2",
    "twine>=5.1",
    "ipykernel>=6.29"
]

[project.scripts]
finance_research_data = "finance_research_data.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/finance_research_data"]
